+++ spring-beans/src/testFixtures/java/org/springframework/beans/testfixture/beans/factory/aot/DeferredTypeBuilder.java
@@ -19,6 +19,7 @@ package org.springframework.beans.testfixture.beans.factory.aot;
 import java.util.function.Consumer;
 
 import org.springframework.javapoet.TypeSpec;
+import org.springframework.lang.Nullable;
 import org.springframework.util.Assert;
 
 /**
@@ -30,6 +31,7 @@ import org.springframework.util.Assert;
  */
 public class DeferredTypeBuilder implements Consumer<TypeSpec.Builder> {
 
+	@Nullable
 	private Consumer<TypeSpec.Builder> type;
 
 	@Override
